+++ frontend/src/components/Metrics/ConfigStep/MetricsTypeCheckbox/index.tsx
