+++ frontend/src/components/Metrics/ConfigStep/MetricsTypeCheckbox/style.tsx
